+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Proving of validity of message field after update with valid sequence (eng/recordflux/RecordFlux#1444)
+
 ## [0.15.0] - 2023-11-08
 
 ### Added
@@ -418,6 +424,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2019-05-14
 
+[Unreleased]: https://github.com/AdaCore/RecordFlux/compare/v0.15.0...HEAD
 [0.15.0]: https://github.com/AdaCore/RecordFlux/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/AdaCore/RecordFlux/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/AdaCore/RecordFlux/compare/v0.12.0...v0.13.0